Fix reacher bug where motors idle. Eliminate need to manually start ray. (#23)
Fixed the bug where the reacher arms would not move. The origin of the issue was that there were some stray commas that made MAX_CURRENT = 0. 

Ray is not started within ars_train so you do not have to run ray start --head before training. This also fixes the issue of the redis IP address sometimes being incorrect.